Feature swc 125 refactorizar iniciar partida

+++ src/components/Partida/Partida.jsx
@@ -1,13 +1,13 @@
 import React, { useContext, useEffect, useState } from "react";
 import { PartidaContext, PartidaProvider } from './PartidaProvider.jsx';
-import Jugador from "./jugador/jugador.jsx";
+import Jugador from "./Jugador/Jugador.jsx";
 import { CartasMovimientoMano } from "./CartasMovimiento/CartasMovimientoMano.jsx";
 import TableroWithProvider from "./Tablero/TableroContainer.jsx";
 import MazoCartaFigura from "./CartaFigura/MazoCartaFigura.jsx";
 import IniciarPartida from "./IniciarPartida/IniciarPartida.jsx";
 import AbandonarPartida from "./AbandonarPartida/AbandonarPartida.jsx";
 import PasarTurno from "./PasarTurno/PasarTurno.jsx";
-import Temporizador from "./temporizador/temporizador.jsx";
+import Temporizador from "./Temporizador/Temporizador.jsx";
 import { WebSocketContext } from '../WebSocketsProvider.jsx';
 import "./Partida.css";
 
@@ -27,17 +27,21 @@ function Partida() {
     isOverlayVisible,
     partidaId
   } = useContext(PartidaContext);
-  
+
   useEffect(() => {
     const jugadoresParseados = JSON.parse(sessionStorage.getItem("players"));
-    if (jugadoresParseados) {
+    if (jugadoresParseados && Array.isArray(jugadoresParseados)) {
       setJugadores(jugadoresParseados);
+      console.log("Jugadores actuales:", jugadoresParseados); 
+    } else {
+      setJugadores([]);  // Asegura que jugadores sea un array vacío
     }
+  
     const partidaIniciada = sessionStorage.getItem("partidaIniciada");
     if (partidaIniciada) {
       setPartidaIniciada(true);
     }
-  }, [sessionStorage.getItem("players"), sessionStorage.getItem("partidaIniciada")]);
+  }, []);
 
   const [timeLeft, setTimeLeft] = useState(() => {
     const storedTime = sessionStorage.getItem("timeLeft");
@@ -45,29 +49,16 @@ function Partida() {
   });
 
   const manejarFinTurno = async () => {
-    const nuevoIndex = (posicionJugador + 1) % jugadores.length;
-    setJugadorActualIndex(nuevoIndex);
-    sessionStorage.setItem("posicion_jugador", nuevoIndex);
-    setTimeLeft(tiempoLimite);
-    sessionStorage.setItem("timeLeft", tiempoLimite);
+    if (jugadores.length > 0) {  // Asegura que jugadores exista antes de acceder
+      const nuevoIndex = (posicionJugador + 1) % jugadores.length;
+      setJugadorActualIndex(nuevoIndex);
+      setPosicionJugador(nuevoIndex);
+      sessionStorage.setItem("posicion_jugador", nuevoIndex);
+      setTimeLeft(tiempoLimite);
+      sessionStorage.setItem("timeLeft", tiempoLimite);
 
-    try {
-      const response = await fetch("https://httpbin.org/post", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ posicionJugador: nuevoIndex }),
-      });
-
-      if (!response.ok) {
-        throw new Error("Error al actualizar el turno en el servidor");
-      }
-    } catch (error) {
-      console.error("Error en la conexión al servidor:", error);
+      setJugando(false);
     }
-
-    setJugando(false);
   };
 
   useEffect(() => {
@@ -76,7 +67,6 @@ function Partida() {
 
   const { wsUPRef } = useContext(WebSocketContext);
 
-  // Escucha el WebSocket de Unirse a Partida
   useEffect(() => {
     try {
       wsUPRef.current.onopen = () => {
@@ -84,9 +74,11 @@ function Partida() {
       };
       wsUPRef.current.onmessage = (event) => {
         const data = JSON.parse(event.data);
-        setJugadores(data.players);
-        sessionStorage.setItem("players", JSON.stringify(data.players || []));
-        console.log("Jugadores recibidos del WebSocket de Unirse a Partida");
+        if (data.players && Array.isArray(data.players)) {
+          setJugadores(data.players);
+          sessionStorage.setItem("players", JSON.stringify(data.players || []));
+          console.log("Jugadores recibidos del WebSocket de Unirse a Partida");
+        }
       };
       wsUPRef.current.onerror = (error) => {
         console.error("Error en el WebSocket de Unirse a Partida:", error);
@@ -105,75 +97,72 @@ function Partida() {
   }, [wsUPRef.current]);
 
   function empezarPartida() {
-    // Reordena jugadores
-
-    // Devuelve la posición del jugador principal o sino -1
     const jugadores = JSON.parse(sessionStorage.getItem("players"));
-    const posicionJugador = jugadores.findIndex(jugador => jugador.id === parseInt(sessionStorage.getItem("player_id"), 10));
-    setPosicionJugador(posicionJugador);
-    if (posicionJugador == -1) {
-      console.log("Jugador no encontrado");
-      sessionStorage.removeItem("players");
-    }
-    else {
-      // Coloca al jugador principal en la primera posición para que se renderice correctamente
-      let jugadoresAux = jugadores[0];
-      jugadores[0] = jugadores[posicionJugador];
-      jugadores[posicionJugador] = jugadoresAux;
-      sessionStorage.setItem("players", JSON.stringify(jugadores));
-      sessionStorage.setItem('partidaIniciada', "true");
-      setPartidaIniciada(true);
-      console.log("Partida iniciada");
+    if (jugadores && Array.isArray(jugadores)) {
+      const posicionJugador = jugadores.findIndex(jugador => jugador.id === parseInt(sessionStorage.getItem("player_id"), 10));
+      setPosicionJugador(posicionJugador);
+      if (posicionJugador !== -1) {
+        let jugadoresAux = jugadores[0];
+        jugadores[0] = jugadores[posicionJugador];
+        jugadores[posicionJugador] = jugadoresAux;
+        sessionStorage.setItem("players", JSON.stringify(jugadores));
+        sessionStorage.setItem('partidaIniciada', "true");
+        setPartidaIniciada(true);
+        console.log("Partida iniciada");
+      } else {
+        console.log("Jugador no encontrado");
+        sessionStorage.removeItem("players");
+      }
     }
   }
 
   return (
     <div className="container-partida">
-      {jugadores && jugadores.map((jugador, index) => (
-        <div key={jugador.id}>
-          <Jugador 
-            nombre={jugador.name} 
-            ubicacion={`jugador${index + 1}`}  
-          />
-          <CartasMovimientoMano
-            ubicacion={index}
-          />
-          <MazoCartaFigura
-            ubicacion={index}
-          />
-        </div>
-      ))}
+      {Array.isArray(jugadores) && jugadores.length > 0 ? (
+        jugadores.map((jugador, index) => (
+          <div key={jugador.id}>
+            <Jugador 
+              nombre={jugador.name} 
+              ubicacion={`jugador${index + 1}`}  
+            />
+            <CartasMovimientoMano ubicacion={index} />
+            <MazoCartaFigura ubicacion={index} />
+          </div>
+        ))
+      ) : (
+        <p>No hay jugadores en la partida.</p>
+      )}
       <div className="tableroContainer">
         <TableroWithProvider />
       </div>
       <div>
         {!partidaIniciada && <IniciarPartida empezarPartida={empezarPartida} />}
       </div>
-      <div className={"abandonar-partida-container"}>
-      {isOverlayVisible && <div className="overlay-supremo"></div>}
+      <div className="abandonar-partida-container">
+        {isOverlayVisible && <div className="overlay-supremo"></div>}
         <AbandonarPartida />
       </div>
       <div className="pasar-turno-container">
         <PasarTurno
-          //jugadorActual={jugadores[jugadorActualIndex]}
+          jugadorActual={jugadores[jugadorActualIndex]}
           jugadores={jugadores}
           onTurnoCambiado={manejarFinTurno}
           tiempoLimite={tiempoLimite}
           setTimeLeft={setTimeLeft}
-          partidaId={partidaId} // Pasar partidaId al componente PasarTurno
+          partidaId={partidaId}
         />
       </div>
-      {/*
       <div className="timer-container">
-        <Temporizador
-          tiempoLimite={tiempoLimite}
-          jugadorActual={jugadores[jugadorActualIndex].name}
-          timeLeft={timeLeft} // Pasar el tiempo restante al Temporizador
-          setTimeLeft={setTimeLeft} // Pasar la función para actualizar el tiempo
-          onFinTurno={manejarFinTurno}
-        />
+        {jugadores && jugadores.length > 0 && (
+          <Temporizador
+            tiempoLimite={tiempoLimite}
+            jugadorActual={jugadores[jugadorActualIndex]?.name}
+            timeLeft={timeLeft}
+            setTimeLeft={setTimeLeft}
+            onFinTurno={manejarFinTurno}
+          />
+        )}
       </div>
-      */}
       {isOverlayVisible && <div className="overlay"></div>}
     </div>
   );
